Add files via upload
For this project involving the transformation and cleaning of the dataset, I chose to showcase my Excel capabilities.

Driving questions:
Are all the data types for each feature consistent?
Are there any missing/unreadable values?
Are there any columns that need to be separated? Combined?
Are there any transformations that need to occur for future comprehension?